+++ src/addon_doctor/__init__.py
@@ -1,3 +1,3 @@
 """Addon Doctor for ESO package."""
 
-__version__ = "0.1.0"
+__version__ = "0.1.1"